FREEZE_Mithos_v1 (#1314)
- Is1273/groups api (#1276)
- Hints + PROGREssbars (son del PSOE) + Favicon + Buttons visibility + filters (#1287)
- Is1289/mail environs (#1291)
- [bug] Fixes database blockage upon many requests on projects interface (#1301)
- Increase timeout for test (#1299)
- Revert "webserver does not respond when many logs are coming in from rabbitMQ (#1293)" (#1297)
- Appends more information in e2e's test screenshots (III) (#1296)
- Appends more information in e2e's test screenshots (II) (#1295)
- more info (#1294)
- webserver does not respond when many logs are coming in from rabbitMQ (#1293)
- Is1233/auto update postgres (#1290)

+++ services/web/server/src/simcore_service_webserver/projects/projects_handlers.py
@@ -202,17 +202,6 @@ async def replace_project(request: web.Request):
 
     return {'data': new_project}
 
-
-# TODO: temporary hidden until get_handlers_from_namespace refactor to seek marked functions instead!
-#@login_required
-#async def patch_project(_request: web.Request):
-#    """
-#        Client sends a patch and return updated project
-#        PATCH
-#    """
-#    # TODO: implement patch with diff as body!
-#    raise NotImplementedError()
-
 @login_required
 async def delete_project(request: web.Request):
     # TODO: replace by decorator since it checks again authentication
@@ -389,6 +378,7 @@ async def delete_node(request: web.Request) -> web.Response:
 
 @login_required
 async def add_tag(request: web.Request):
+    await check_permission(request, 'project.tag.*')
     uid, db = request[RQT_USERID_KEY], request.config_dict[APP_PROJECT_DBAPI]
     tag_id, study_uuid = request.match_info.get('tag_id'), request.match_info.get('study_uuid')
     return await db.add_tag(
@@ -400,6 +390,7 @@ async def add_tag(request: web.Request):
 
 @login_required
 async def remove_tag(request: web.Request):
+    await check_permission(request, 'project.tag.*')
     uid, db = request[RQT_USERID_KEY], request.config_dict[APP_PROJECT_DBAPI]
     tag_id, study_uuid = request.match_info.get('tag_id'), request.match_info.get('study_uuid')
     return await db.remove_tag(
